% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/eu_data.R
\docType{data}
\name{eu}
\alias{eu}
\title{Testate data for European transfer function}
\format{
An object of class \code{data.frame} with 1462 rows and 56 columns.

An object of class \code{data.frame} with 1462 rows and 56 columns.
}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0277379116303742#appsec1}

\href{https://www.sciencedirect.com/science/article/pii/S0277379116303742#appsec1}
}
\usage{
data(eu)

#' @keywords datasets

data(eu)

#' @keywords datasets
}
\description{
Data contains testates abundancies alongside pH and WTD
from countries across Europe allowing for the creation of
a transfer function

Data contains testates abundancies alongside pH and WTD
from countries across Europe allowing for the creation of
a transfer function
}
\examples{
data(eu)
times <- attr(eu, "time")
data(eu)
}
\references{
Amesbury et al. (2016)
https://doi.org/10.1016/j.quascirev.2016.09.024

Amesbury et al. (2016)
"Development of a new pan-European testate amoeba transfer
 function for reconstructing peatland palaeohydrology"
https://doi.org/10.1016/j.quascirev.2016.09.024
}
\keyword{datasets}
