% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_data.R
\docType{data}
\name{na}
\alias{na}
\title{Testate data for a North American transfer function}
\format{
An object of class \code{data.frame} with 1907 rows and 80 columns.
}
\source{
\href{https://www.sciencedirect.com/science/article/pii/S0277379118304530#ec-research-data}
}
\usage{
data(na)
}
\description{
Data contains testates abundances alongside pH and WTD
from the USA and Canada allowing for the creation of
a transfer function
}
\examples{
data(na)
}
\references{
Amesbury et al. (2018)
"Towards a Holarctic synthesis of peatland testate amoeba ecology:
Development of a new continental-scale palaeohydrological
transfer function for North America and comparison to European data"
https://doi.org/10.1016/j.quascirev.2018.10.034
}
\keyword{datasets}
