% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transfer function.R
\name{transfer}
\alias{transfer}
\title{Create, apply and graph the results of a transfer function}
\usage{
transfer(
  x = "lh1_tests",
  tf = "eu",
  country = "england",
  save = T,
  age_file = "lh1_ages",
  depth_start = 1,
  depth_int = 1,
  boot_size = 1000
)
}
\arguments{
\item{x}{Name of your own data to apply the transfer function to}

\item{tf}{The continent for which you want the transfer function data for (na or eu)}

\item{country}{The country within the continental data you wish to choose}

\item{save}{Choose to save the output to disk}

\item{age_file}{The name of the optional age file to graph the results with}

\item{depth_start}{The first depth you sampled}

\item{depth_int}{The interval in cm between depths of each sample}

\item{boot_size}{The numbe of bootstrap cycles to run}
}
\value{
Performance of the model, reconstruction, zscores and error, graphs of WTD and Zscores
}
\description{
Create, apply and graph the results of a transfer function
}
\examples{
transfer(x = "lh1_tests", tf = "eu", country = "england", save = T, age_file = "lh1_ages", depth_start = 1, depth_int = 1, boot_size = 1000)
}
